// app/packages/graphin/src/types.ts
export type GraphClass = typeof G6['Graph'];
export type Graph = InstanceType<GraphClass>;
type GraphOptions = ConstructorParameters<GraphClass>[0];
export type GraphNode = NodeConfig;
export type GraphEdge = EdgeConfig;

/** G6 没有暴露这个类型 */
export interface G6Event extends MouseEvent {
  item: Item;
  target: MouseEvent['target'];
}

export interface G6KeyboardEvent extends KeyboardEvent {
  item: Item;
  target: KeyboardEvent['target'];
}

/** Graphin 用到一些 G6 官方没有暴露的 API，所以需要 extend 官方的类型 */
export interface ExtendedGraphOptions extends GraphOptions {
  /**
   * 默认缩放比例
   */
  zoom?: number;

  pan?: { x: number; y: number };

  disablePan?: boolean; // 禁止画布平移
  disableZoom?: boolean; // 禁用画布缩放
  disableDrag?: boolean; // 禁用节点拖拽
  disableHighlight?: boolean; // 禁用highlight
  delegateNode?: boolean;
  wheelSensitivity?: number;
  isZoomOptimize?: (graph?: Graph, e?: G6Event) => boolean;
  keyShapeZoom?: number;
  autoFollowWithForce?: boolean;

  /**
   * 开启多边支持，默认否
   *
   * @type {boolean}
   * @memberof ExtendedGraphOptions
   */
  autoPolyEdge: boolean;

  /**
   * 开启自环边支持，默认是
   *
   * @type {boolean}
   * @memberof ExtendedGraphOptions
   */
  autoLoopEdge: boolean;

  autoPinWithForce?: boolean;
  restartForceOnDrag?: boolean;
  [key: string]: any; // eslint-disable-line
}
// ...
export interface Node {
  /** 节点源数据 */
  data: {
    /** 唯一标示ID，必选 */
    id: string;
    /** comboId */
    comboId?: string;
    /** 节点数据类型 */
    type?: string;
    /** 节点文本 */
    label?: string;
    /** 节点属性 */
    properties?: any[]; // eslint-disable-line
  };
  /** 唯一标示ID，必选 */
  id: string;
  /** comboId */
  comboId?: string;
  /** 节点类型 */
  shape?: string;
  /** 节点类型 */
  type?: string;
  /** 节点文本 */
  label?: string;
  /** 节点样式 */
  style?: Partial<NodeStyle>;
  /** 内置节点属性 */
  // innerNodeStyle?: Partial<InnerNodeStyle>;
  /** 节点位置信息 */
  x?: number;
  y?: number;
  /** 内置的布局，追加的额外属性 */
  layout?: NodeLayoutType;
  /** 节点度数 */
  degree?: number;
  /** 徽标文案 */
  badge?: number | string;
  searchTypes?: string[];

  /** User Defined Propery */
  [key: string]: any; // eslint-disable-line @typescript-eslint/no-explicit-any
}
// ...
export interface Edge {
  /** 边的数据，必选 */
  data: {
    /** 边的属性 */
    properties?: object[];

    // eslint-disable-next-line @typescript-eslint/no-explicit-any
    [key: string]: any;
  };
  /** 边的源节点，必选 */
  source: string;
  /** 边的目标节点，必选 */
  target: string;
  /** 边的类型 */
  shape?: string;
  /** 边的类型 */
  type?: string;

  /** 边的文本 */
  label?: string;
  /** 边的样式 */
  style?: InnerEdgeStyle;

  id?: string;
  /** 边的弹簧长度，力导时使用 */
  spring?: number;

  /**
   * 自环
   *
   * @type {({
   *     position: string,
   *     dist: number,
   *   })}
   * @memberof Edge
   */
  loopCfg?: {
    position: string;
    dist: number;
  };

  /**
   * 多边
   *
   * @type {[number, number][]}
   * @memberof Edge
   */
  poly?: {
    distance: number;
  };

  /** User Defined Property */
  [key: string]: any; // eslint-disable-line @typescript-eslint/no-explicit-any
}

export type G6Edge = Edge & {
  startPoint: G6Node;
  endPoint: G6Node;
  sourceNode: G6ItemNode;
  targetNode: G6ItemNode;
};

export type NodeData = Node['data'];
export type EdgeData = Edge['data'];

export interface Data {
  /** 节点 */
  nodes: Node[];
  /** 边 */
  edges: Edge[];
  /** combo */
  combos?: ComboConfig[];
}
export interface Layout {
  /** 布局名称，必选 */
  name: string;
  /** 布局配置，可选 */
  options?: {
    [key: string]: any; // eslint-disable-line
  };
}
// ...
export interface NodeShapeFunction {
  (node: Node): {
    /** 自定义Shape的名称，之后数据指定即可调用这个Shape定义 */
    shape?: string;
    type?: string;
    shapeComponents: ShapeComponent[];
    state: {
      [stateName: string]: {
        [id: string]: {
          [attr: string]: any; // eslint-disable-line
        };
      };
    };
  };
}

export interface ExtendNodeShape {
  name: string;
  render: NodeShapeFunction;
}
export interface ExtendLayout {
  /** 布局名称，唯一标示 */
  name: string;
  /** 布局展示名称 */
  desc: string;
  /** antd icon */
  icon: string;
  /** layout 布局执行函数 */
  layout: (
    data: Data,
    options: LayoutOption
  ) => {
    data: Data;
    forceSimulation?: ForceSimulation;
  };
}
export interface ExtendMarker {
  /** 唯一标示名称 */
  name: string;
  /** Marker的Path路径 */
  path: string;
}

export interface IconFontMapItem {
  name: string;
  unicode_decimal: number;
}

export interface ExtendIcon {
  /** 字体名称 */
  fontFamily: string;
  /** iconfont 的 class name 和 unicode decimal 的映射 */
  map: IconFontMapItem[];
}

export interface Register {
  /** 节点名称 */
  name: string;
  /** register执行函数,参数为G6对象 */
  register: (g6: typeof G6) => void;
}

export interface BehaviorRegister extends Register {
  options: any; // eslint-disable-line
  mode: string;
}

export interface GraphinProps {
  /** render */
  data: Data;
  /** options */
  options?: Partial<ExtendedGraphOptions>;
  /** layout */
  layout?: Layout;

  extend?: {
    layout?: (graphin: Graphin, prevProps: GraphinProps) => ExtendLayout[];
    nodeShape?: () => ExtendNodeShape[];
    marker?: () => ExtendMarker[];
    icon?: () => ExtendIcon[];
  };
  register?: {
    /** 通过G6原生方法，注册节点 */
    nodeShape?: (g6: typeof G6) => Register[];
    /** 通过G6原生方法，注册边 */
    edgeShape?: (g6: typeof G6) => Register[];
    /** 通过G6原生方法，注册事件 */
    behavior?: (g6: typeof G6) => BehaviorRegister[];
  };

  children?: ReactNode;

  [key: string]: GraphinProps[keyof GraphinProps];
}

export type ForceSimulation = ForceLayout;